+++ index.test.js
@@ -14,7 +14,7 @@ beforeEach(() => {
   delete document.documentElement.dataset.tnaCookiesDomain;
   delete document.documentElement.dataset.tnaCookiesPath;
   delete document.documentElement.dataset.tnaCookiesSecure;
-  delete document.documentElement.dataset.tnaCookiesPoliciesKey;
+  delete document.documentElement.dataset.tnaCookiesPreferencesKey;
   delete document.documentElement.dataset.tnaCookiesDefaultAge;
 });
 
@@ -183,8 +183,8 @@ describe("IIFE", () => {
 
     expect(window.TNAFrontendCookies).not.toBeFalsy();
 
-    const { policiesCorrectOnInit1, ...cookies1Properties } = cookies1;
-    const { policiesCorrectOnInit2, ...cookies2Properties } = cookies2;
+    const { preferencesCorrectOnInit1, ...cookies1Properties } = cookies1;
+    const { preferencesCorrectOnInit2, ...cookies2Properties } = cookies2;
     expect(cookies1Properties).toEqual(cookies2Properties);
     expect(cookies2).toHaveProperty("customProperty");
   });
@@ -335,36 +335,36 @@ describe("Events", () => {
     cookies.on("deleteCookie", mockDeleteCookieCallback);
     const mockDeleteAllCookiesCallback = vi.fn();
     cookies.on("deleteAllCookies", mockDeleteAllCookiesCallback);
-    const mockAcceptPolicyCallback = vi.fn();
-    cookies.on("acceptPolicy", mockAcceptPolicyCallback);
-    const mockRejectPolicyCallback = vi.fn();
-    cookies.on("rejectPolicy", mockRejectPolicyCallback);
-    const mockAcceptAllPoliciesCallback = vi.fn();
-    cookies.on("acceptAllPolicies", mockAcceptAllPoliciesCallback);
-    const mockRejectAllPoliciesCallback = vi.fn();
-    cookies.on("rejectAllPolicies", mockRejectAllPoliciesCallback);
-    const mockChangePolicyCallback = vi.fn();
-    cookies.on("changePolicy", mockChangePolicyCallback);
+    const mockEnablePreferenceCallback = vi.fn();
+    cookies.on("enablePreference", mockEnablePreferenceCallback);
+    const mockDisablePreferenceCallback = vi.fn();
+    cookies.on("disablePreference", mockDisablePreferenceCallback);
+    const mockEnsableAllPreferencesCallback = vi.fn();
+    cookies.on("ensableAllPreferences", mockEnsableAllPreferencesCallback);
+    const mockDisableAllPreferencesCallback = vi.fn();
+    cookies.on("disableAllPreferences", mockDisableAllPreferencesCallback);
+    const mockChangePreferenceCallback = vi.fn();
+    cookies.on("changePreference", mockChangePreferenceCallback);
 
     const testKey = "foo";
     const testValue = "bar";
     cookies.set(testKey, testValue);
     cookies.delete(testKey);
-    cookies.acceptPolicy("settings");
-    cookies.rejectPolicy("settings");
-    cookies.setPolicy("settings", true);
-    cookies.acceptAllPolicies();
-    cookies.rejectAllPolicies();
+    cookies.enablePreference("settings");
+    cookies.disablePreference("settings");
+    cookies.setPreference("settings", true);
+    cookies.ensableAllPreferences();
+    cookies.disableAllPreferences();
     cookies.deleteAll();
 
     expect(mockSetCookieCallback.mock.calls).toHaveLength(8);
     expect(mockDeleteCookieCallback.mock.calls).toHaveLength(2);
     expect(mockDeleteAllCookiesCallback.mock.calls).toHaveLength(1);
-    expect(mockAcceptPolicyCallback.mock.calls).toHaveLength(1);
-    expect(mockRejectPolicyCallback.mock.calls).toHaveLength(1);
-    expect(mockAcceptAllPoliciesCallback.mock.calls).toHaveLength(1);
-    expect(mockRejectAllPoliciesCallback.mock.calls).toHaveLength(1);
-    expect(mockChangePolicyCallback.mock.calls).toHaveLength(7);
+    expect(mockEnablePreferenceCallback.mock.calls).toHaveLength(1);
+    expect(mockDisablePreferenceCallback.mock.calls).toHaveLength(1);
+    expect(mockEnsableAllPreferencesCallback.mock.calls).toHaveLength(1);
+    expect(mockDisableAllPreferencesCallback.mock.calls).toHaveLength(1);
+    expect(mockChangePreferenceCallback.mock.calls).toHaveLength(7);
   });
 
   test("One-time events", async () => {
@@ -419,21 +419,21 @@ describe("Initialisation", () => {
     const cookies = new Cookies({
       defaultDomain: ".localhost",
       secure: true,
-      policiesKey: "cookies_policy_custom",
+      preferencesKey: "cookie_preferences_custom",
       defaultAge: 300,
     });
     expect(cookies.defaultDomain).toEqual(".localhost");
     expect(cookies.secure).toEqual(true);
-    expect(cookies.policiesKey).toEqual("cookies_policy_custom");
+    expect(cookies.preferencesKey).toEqual("cookie_preferences_custom");
     expect(cookies.defaultAge).toEqual(300);
 
     expect(document.cookie).not.toEqual("");
 
-    expect(cookies.exists("cookies_policy_custom")).toEqual(true);
-    cookies.acceptAllPolicies();
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
+    expect(cookies.exists("cookie_preferences_custom")).toEqual(true);
+    cookies.ensableAllPreferences();
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
   });
 
   test("With a different path", async () => {
@@ -447,8 +447,8 @@ describe("Initialisation", () => {
 
   test("Properties from HTML attributes", async () => {
     document.documentElement.dataset.tnaCookiesDomain = ".localhost";
-    document.documentElement.dataset.tnaCookiesPoliciesKey =
-      "cookies_policy_custom";
+    document.documentElement.dataset.tnaCookiesPreferencesKey =
+      "cookie_preferences_custom";
     document.documentElement.dataset.tnaCookiesDefaultAge = 300;
     document.documentElement.dataset.tnaCookiesInsecure = true;
 
@@ -456,353 +456,355 @@ describe("Initialisation", () => {
 
     expect(cookies.defaultDomain).toEqual(".localhost");
     expect(cookies.secure).toEqual(false);
-    expect(cookies.policiesKey).toEqual("cookies_policy_custom");
+    expect(cookies.preferencesKey).toEqual("cookie_preferences_custom");
     expect(cookies.defaultAge).toEqual(300);
   });
 
   test("Explicit properties overwrite HTML attributes", async () => {
     document.documentElement.dataset.tnaCookiesDomain = ".localhost";
-    document.documentElement.dataset.tnaCookiesPoliciesKey =
-      "cookies_policy_custom";
+    document.documentElement.dataset.tnaCookiesPreferencesKey =
+      "cookie_preferences_custom";
     document.documentElement.dataset.tnaCookiesDefaultAge = 300;
     document.documentElement.dataset.tnaCookiesInsecure = true;
 
     const cookies = new Cookies({
       defaultDomain: "localhost",
-      policiesKey: "cookies_policy_custom2",
+      preferencesKey: "cookie_preferences_custom2",
       secure: true,
       defaultAge: 600,
     });
 
     expect(cookies.defaultDomain).toEqual("localhost");
     expect(cookies.secure).toEqual(true);
-    expect(cookies.policiesKey).toEqual("cookies_policy_custom2");
+    expect(cookies.preferencesKey).toEqual("cookie_preferences_custom2");
     expect(cookies.defaultAge).toEqual(600);
   });
 });
 
-describe("Policies", () => {
+describe("Preferences", () => {
   test("Default preferences on initialisation", async () => {
     const cookies = new Cookies();
 
-    expect(cookies).toHaveProperty("policies");
+    expect(cookies).toHaveProperty("preferences");
 
-    expect(cookies.all).toHaveProperty("cookies_policy");
-    expect(cookies.exists("cookies_policy")).toEqual(true);
+    expect(cookies.all).toHaveProperty("cookie_preferences");
+    expect(cookies.exists("cookie_preferences")).toEqual(true);
 
-    expect(cookies.policies).toHaveProperty("essential");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(JSON.parse(cookies.get("cookies_policy"))).toHaveProperty(
+    expect(cookies.preferences).toHaveProperty("essential");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(JSON.parse(cookies.get("cookie_preferences"))).toHaveProperty(
       "essential",
     );
-    expect(cookies.policies).toHaveProperty("settings");
-    expect(cookies.policies.settings).toEqual(false);
-    expect(JSON.parse(cookies.get("cookies_policy"))).toHaveProperty(
+    expect(cookies.preferences).toHaveProperty("settings");
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(JSON.parse(cookies.get("cookie_preferences"))).toHaveProperty(
       "settings",
     );
-    expect(cookies.policies).toHaveProperty("usage");
-    expect(cookies.policies.usage).toEqual(false);
-    expect(JSON.parse(cookies.get("cookies_policy"))).toHaveProperty("usage");
-    expect(cookies.policies).toHaveProperty("marketing");
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(JSON.parse(cookies.get("cookies_policy"))).toHaveProperty(
+    expect(cookies.preferences).toHaveProperty("usage");
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(JSON.parse(cookies.get("cookie_preferences"))).toHaveProperty(
+      "usage",
+    );
+    expect(cookies.preferences).toHaveProperty("marketing");
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(JSON.parse(cookies.get("cookie_preferences"))).toHaveProperty(
       "marketing",
     );
   });
 
-  test("Get non-existent policy", async () => {
+  test("Get non-existent preference", async () => {
     const cookies = new Cookies();
 
-    expect(() => cookies.isPolicyAccepted("foobar")).toThrow(
-      new Error("Policy 'foobar' does not exist"),
+    expect(() => cookies.isPreferenceAccepted("foobar")).toThrow(
+      new Error("Preference 'foobar' does not exist"),
     );
   });
 
-  test("Accept policy", async () => {
+  test("Accept preference", async () => {
     const cookies = new Cookies();
-    expect(cookies).toHaveProperty("isPolicyAccepted");
+    expect(cookies).toHaveProperty("isPreferenceAccepted");
 
-    expect(cookies.policies).toHaveProperty("essential");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies).toHaveProperty("settings");
-    expect(cookies.policies.settings).toEqual(false);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-    expect(cookies.policies).toHaveProperty("usage");
-    expect(cookies.policies.usage).toEqual(false);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-    expect(cookies.policies).toHaveProperty("marketing");
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+    expect(cookies.preferences).toHaveProperty("essential");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences).toHaveProperty("settings");
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+    expect(cookies.preferences).toHaveProperty("usage");
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+    expect(cookies.preferences).toHaveProperty("marketing");
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
 
-    cookies.acceptPolicy("settings");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(true);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.policies.usage).toEqual(false);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+    cookies.enablePreference("settings");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(true);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
 
-    cookies.acceptPolicy("usage");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(true);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.policies.usage).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+    cookies.enablePreference("usage");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(true);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.preferences.usage).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
 
-    cookies.acceptPolicy("marketing");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(true);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.policies.usage).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.policies.marketing).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
+    cookies.enablePreference("marketing");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(true);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.preferences.usage).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.preferences.marketing).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
   });
 
-  test("Accept non-existent policy", async () => {
+  test("Accept non-existent preference", async () => {
     const cookies = new Cookies();
 
-    expect(() => cookies.acceptPolicy("foobar")).toThrow(Error);
+    expect(() => cookies.enablePreference("foobar")).toThrow(Error);
   });
 
-  test("Reject non-existent policy", async () => {
+  test("Reject non-existent preference", async () => {
     const cookies = new Cookies();
 
-    expect(() => cookies.rejectPolicy("foobar")).toThrow(Error);
+    expect(() => cookies.disablePreference("foobar")).toThrow(Error);
   });
 
-  test("Set non-existent policy", async () => {
+  test("Set non-existent preference", async () => {
     const cookies = new Cookies();
 
-    expect(() => cookies.setPolicy("foobar", true)).toThrow(Error);
+    expect(() => cookies.setPreference("foobar", true)).toThrow(Error);
   });
 
-  test("Accept all policies", async () => {
+  test("Accept all preferences", async () => {
     const cookies = new Cookies();
-    expect(cookies).toHaveProperty("acceptAllPolicies");
+    expect(cookies).toHaveProperty("ensableAllPreferences");
 
-    expect(cookies.policies).toHaveProperty("essential");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies).toHaveProperty("settings");
-    expect(cookies.policies.settings).toEqual(false);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-    expect(cookies.policies).toHaveProperty("usage");
-    expect(cookies.policies.usage).toEqual(false);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-    expect(cookies.policies).toHaveProperty("marketing");
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+    expect(cookies.preferences).toHaveProperty("essential");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences).toHaveProperty("settings");
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+    expect(cookies.preferences).toHaveProperty("usage");
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+    expect(cookies.preferences).toHaveProperty("marketing");
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
 
-    cookies.acceptAllPolicies();
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(true);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.policies.usage).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.policies.marketing).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
+    cookies.ensableAllPreferences();
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(true);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.preferences.usage).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.preferences.marketing).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
   });
 
-  test("Reject policy", async () => {
-    const cookies = new Cookies();
-
-    cookies.acceptPolicy("settings");
-    cookies.acceptPolicy("usage");
-    cookies.acceptPolicy("marketing");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(true);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.policies.usage).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.policies.marketing).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
-
-    cookies.rejectPolicy("settings");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(false);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-    expect(cookies.policies.usage).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.policies.marketing).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
-
-    cookies.rejectPolicy("usage");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(false);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-    expect(cookies.policies.usage).toEqual(false);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-    expect(cookies.policies.marketing).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
-
-    cookies.rejectPolicy("marketing");
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(false);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-    expect(cookies.policies.usage).toEqual(false);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
-  });
-
-  test("Reject all policies", async () => {
-    const cookies = new Cookies();
-    expect(cookies).toHaveProperty("rejectAllPolicies");
-
-    cookies.acceptAllPolicies();
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(true);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-    expect(cookies.policies.usage).toEqual(true);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-    expect(cookies.policies.marketing).toEqual(true);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
-
-    cookies.rejectAllPolicies();
-    expect(cookies.policies.essential).toEqual(true);
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-    expect(cookies.policies.settings).toEqual(false);
-    expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-    expect(cookies.policies.usage).toEqual(false);
-    expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-    expect(cookies.policies.marketing).toEqual(false);
-    expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
-  });
-
-  test("Protected essential policy", async () => {
+  test("Reject preference", async () => {
     const cookies = new Cookies();
 
-    expect(cookies.policies).toHaveProperty("essential");
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+    cookies.enablePreference("settings");
+    cookies.enablePreference("usage");
+    cookies.enablePreference("marketing");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(true);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.preferences.usage).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.preferences.marketing).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
 
-    cookies.acceptPolicy("essential");
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+    cookies.disablePreference("settings");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+    expect(cookies.preferences.usage).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.preferences.marketing).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
 
-    cookies.rejectPolicy("essential");
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+    cookies.disablePreference("usage");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+    expect(cookies.preferences.marketing).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
 
-    cookies.rejectAllPolicies();
-    expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+    cookies.disablePreference("marketing");
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
   });
 
-  test("Custom policy key", async () => {
-    const cookies = new Cookies({ policiesKey: ["custom"] });
+  test("Reject all preferences", async () => {
+    const cookies = new Cookies();
+    expect(cookies).toHaveProperty("disableAllPreferences");
 
-    expect(cookies.all).not.toHaveProperty("cookies_policy");
+    cookies.ensableAllPreferences();
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(true);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+    expect(cookies.preferences.usage).toEqual(true);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+    expect(cookies.preferences.marketing).toEqual(true);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
+
+    cookies.disableAllPreferences();
+    expect(cookies.preferences.essential).toEqual(true);
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+    expect(cookies.preferences.settings).toEqual(false);
+    expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+    expect(cookies.preferences.usage).toEqual(false);
+    expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+    expect(cookies.preferences.marketing).toEqual(false);
+    expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
+  });
+
+  test("Protected essential preference", async () => {
+    const cookies = new Cookies();
+
+    expect(cookies.preferences).toHaveProperty("essential");
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+
+    cookies.enablePreference("essential");
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+
+    cookies.disablePreference("essential");
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+
+    cookies.disableAllPreferences();
+    expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+  });
+
+  test("Custom preference key", async () => {
+    const cookies = new Cookies({ preferencesKey: ["custom"] });
+
+    expect(cookies.all).not.toHaveProperty("cookie_preferences");
     expect(cookies.all).toHaveProperty("custom");
 
-    expect(cookies.policies.settings).toEqual(false);
+    expect(cookies.preferences.settings).toEqual(false);
 
-    cookies.acceptAllPolicies();
-    expect(cookies.policies.settings).toEqual(true);
+    cookies.ensableAllPreferences();
+    expect(cookies.preferences.settings).toEqual(true);
   });
 
   describe("With existing cookies", () => {
-    test("Valid policy", async () => {
+    test("Valid preference", async () => {
       document.cookie =
-        "cookies_policy=%7B%22essential%22%3Atrue%2C%22settings%22%3Atrue%2C%22usage%22%3Atrue%2C%22marketing%22%3Atrue%7D";
+        "cookie_preferences=%7B%22essential%22%3Atrue%2C%22settings%22%3Atrue%2C%22usage%22%3Atrue%2C%22marketing%22%3Atrue%7D";
 
       const cookies = new Cookies();
 
-      expect(cookies.policiesCorrectOnInit).toEqual(true);
-      expect(cookies.all).toHaveProperty("cookies_policy");
-      expect(cookies.policies).toHaveProperty("essential");
-      expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("settings");
-      expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("usage");
-      expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("marketing");
-      expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
+      expect(cookies.preferencesCorrectOnInit).toEqual(true);
+      expect(cookies.all).toHaveProperty("cookie_preferences");
+      expect(cookies.preferences).toHaveProperty("essential");
+      expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("settings");
+      expect(cookies.isPreferenceAccepted("settings")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("usage");
+      expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("marketing");
+      expect(cookies.isPreferenceAccepted("marketing")).toEqual(true);
     });
 
-    test("Empty policy", async () => {
-      document.cookie = "cookies_policy=%7B%7D";
+    test("Empty preference", async () => {
+      document.cookie = "cookie_preferences=%7B%7D";
 
       const cookies = new Cookies();
 
-      expect(cookies.policiesCorrectOnInit).toEqual(false);
-      expect(cookies.all).toHaveProperty("cookies_policy");
-      expect(cookies.policies).toHaveProperty("essential");
-      expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("settings");
-      expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("usage");
-      expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("marketing");
-      expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+      expect(cookies.preferencesCorrectOnInit).toEqual(false);
+      expect(cookies.all).toHaveProperty("cookie_preferences");
+      expect(cookies.preferences).toHaveProperty("essential");
+      expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("settings");
+      expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("usage");
+      expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("marketing");
+      expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
     });
 
-    test("Invalid policy", async () => {
-      document.cookie = "cookies_policy=foobar";
+    test("Invalid preference", async () => {
+      document.cookie = "cookie_preferences=foobar";
 
       const cookies = new Cookies();
 
-      expect(cookies.policiesCorrectOnInit).toEqual(false);
-      expect(cookies.all).toHaveProperty("cookies_policy");
-      expect(cookies.policies).toHaveProperty("essential");
-      expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("settings");
-      expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("usage");
-      expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("marketing");
-      expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+      expect(cookies.preferencesCorrectOnInit).toEqual(false);
+      expect(cookies.all).toHaveProperty("cookie_preferences");
+      expect(cookies.preferences).toHaveProperty("essential");
+      expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("settings");
+      expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("usage");
+      expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("marketing");
+      expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
     });
 
-    test("Partial policy", async () => {
-      document.cookie = "cookies_policy=%7B%22usage%22%3Atrue%7D";
+    test("Partial preference", async () => {
+      document.cookie = "cookie_preferences=%7B%22usage%22%3Atrue%7D";
 
       const cookies = new Cookies();
 
-      expect(cookies.policiesCorrectOnInit).toEqual(false);
-      expect(cookies.all).toHaveProperty("cookies_policy");
-      expect(cookies.policies).toHaveProperty("essential");
-      expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("settings");
-      expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("usage");
-      expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("marketing");
-      expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+      expect(cookies.preferencesCorrectOnInit).toEqual(false);
+      expect(cookies.all).toHaveProperty("cookie_preferences");
+      expect(cookies.preferences).toHaveProperty("essential");
+      expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("settings");
+      expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("usage");
+      expect(cookies.isPreferenceAccepted("usage")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("marketing");
+      expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
     });
 
-    test("Unknown policy", async () => {
-      document.cookie = "cookies_policy=%7B%22custom%22%3Atrue%7D";
+    test("Unknown preference", async () => {
+      document.cookie = "cookie_preferences=%7B%22custom%22%3Atrue%7D";
 
       const cookies = new Cookies();
 
-      expect(cookies.policiesCorrectOnInit).toEqual(false);
-      expect(cookies.all).toHaveProperty("cookies_policy");
-      expect(cookies.policies).toHaveProperty("essential");
-      expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-      expect(cookies.policies).toHaveProperty("settings");
-      expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("usage");
-      expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-      expect(cookies.policies).toHaveProperty("marketing");
-      expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+      expect(cookies.preferencesCorrectOnInit).toEqual(false);
+      expect(cookies.all).toHaveProperty("cookie_preferences");
+      expect(cookies.preferences).toHaveProperty("essential");
+      expect(cookies.isPreferenceAccepted("essential")).toEqual(true);
+      expect(cookies.preferences).toHaveProperty("settings");
+      expect(cookies.isPreferenceAccepted("settings")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("usage");
+      expect(cookies.isPreferenceAccepted("usage")).toEqual(false);
+      expect(cookies.preferences).toHaveProperty("marketing");
+      expect(cookies.isPreferenceAccepted("marketing")).toEqual(false);
     });
   });
 
   test("No initialisation", async () => {
     const cookies = new Cookies({ noInit: true });
 
-    expect(cookies.policiesCorrectOnInit).toEqual(false);
-    expect(cookies.all).not.toHaveProperty("cookies_policy");
+    expect(cookies.preferencesCorrectOnInit).toEqual(false);
+    expect(cookies.all).not.toHaveProperty("cookie_preferences");
   });
 });